Add files via upload
CardioRisk Navigator is an offline platform designed and developed to explore the path of machine learning . The goal is to predict the health of a patient from collective data, so as to be able to detect configurations at risk for the patient, and therefore, in cases requiring emergency medical assistance, alert the appropriate medical staff of the situation of the latter.
We initially have a dataset collecting information of many patients with which we are able to conclude the results into a complete form and can predict data precisely. The results of the predictions, derived from the predictive models generated by machine learning, will be presented through several distinct graphical interfaces according to the datasets considered. We will then bring criticism as to the scope of our results.